+++ R/description.R
@@ -43,16 +43,16 @@ description <- function(dataSet, level = 1, path_to_write = NULL, verbose = TRUE
   # Level 1: Univariate description
   if (level >= 1){
     cat("#####################################\n## Level 1: univariate description ##\n#####################################\n")
-	card <- dataSet[, lapply(.SD, uniqueN)]
+    card <- dataSet[, lapply(.SD, uniqueN)]
     for (col in colnames(dataSet)){
-	  # Unique or distinct values
+      # Unique or distinct values
       if (card[[col]] == 1){
-		printl(col, " only has 1 value.")
-	  } 
-	  if (card[[col]] == nrow(dataSet)){
-	    printl(col, " has all unique values.")
-	  }
-	  # Numerical and date cols
+        printl(col, " only has 1 value.")
+      } 
+      if (card[[col]] == nrow(dataSet)){
+        printl(col, " has all unique values.")
+      }
+      # Numerical and date cols
       if (is.numeric(dataSet[[col]]) || is.date(dataSet[[col]])){
         if (is.numeric(dataSet[[col]])){
           printl("Summary for numeric variable ", col)  
@@ -62,7 +62,7 @@ description <- function(dataSet, level = 1, path_to_write = NULL, verbose = TRUE
         }
         print(summary(dataSet[[col]]))
       }
-	  # Factor cols
+      # Factor cols
       if (is.factor(dataSet[[col]]) || is.logical(dataSet[[col]])){
         if (is.factor(dataSet[[col]])){
           printl("Table of occurence for factor variable ", col)  
@@ -72,7 +72,7 @@ description <- function(dataSet, level = 1, path_to_write = NULL, verbose = TRUE
         }
         print(table(dataSet[[col]]))
       }
-	  # character
+      # character
       if (is.character(dataSet[[col]])){
         printl("character variable ", col, " takes ", uniqueN(dataSet[[col]]), " different values.")
       }