+++ tests/testthat/test_fast_posixct.R
@@ -1,0 +1,7 @@
+context("test_fast_dates")
+
+test_that("Function as.POSIXct_fast:",
+          {
+            expect_equal(as.POSIXct_fast("2018-06-12", format="%Y-%m-%d"), as.POSIXct("2018-06-12", format="%Y-%m-%d"))
+          })
+