+++ NEWS.md
@@ -1,3 +1,23 @@
+V 0.3.7
+=======
+- New features:
+	- New functions:
+		- Function *as.POSIXct_fast* is now available. It helps to transform to POSIXct way faster (if the same date value is present multiple times in the column).
+	- New features in existing functions:
+		- In dates identifications, we make it faster by computing search of format only on unique values.
+		- In date transformation, we made it faster by using *as.POSIXct_fast* when it is necessary.
+		- Functions *findAndTransFormDates*, *findAndTransformNumerics* and *unFactor* now accept argument *cols* to limitate search.
+
+- Bug fixes:
+	- Control that over-allocate option is activated on every data.table to avoid issues with set. Package should be more robust.
+	- In bijection search (internal function *fastIsBijection*) there was a bug on some rare cases. Fixed but slower.
+
+-Code quality:
+	- Improving code quality using lintr
+	- Suppressing some useless code
+	- Meeting new covr standard
+	- Improve log of setColAsXXX
+
 V 0.3.6
 =======
 - Bug fixes: